+++ packages/testing/src/execution_testing/forks/forks/eips/cancun/__init__.py
@@ -1,0 +1,1 @@
+"""Listings of all EIPs for Cancun fork."""
